Add AgreeDisagreeComponent and fix SliderComponent
