+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "axios": "^1.6.1",
     "next": "14.0.0",
+    "rc-slider": "^10.4.0",
     "react": "^18",
     "react-dom": "^18"
   },
